partial  refactor
